🚨 fix/refactor: Resolve compile error due to merge conflict and improve code structure
