+++ src/main/java/com/jokim/sivillage/api/bridge/reviewmedialist/infrastructure/ReviewMediaListRepository.java
@@ -4,7 +4,5 @@ import com.jokim.sivillage.api.bridge.reviewmedialist.domain.ReviewMediaList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewMediaListRepository extends JpaRepository<ReviewMediaList, Long> {
-  
-    List<ReviewMediaList> findByReviewCode(String reviewCode);
 
 }
