+++ src/main/java/com/jokim/sivillage/api/media/domain/Media.java
@@ -28,6 +28,6 @@ public class Media {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private MediaType mediaType = MediaType.IMAGE;
+    private MediaType mediaType;
 
 }
